StrativAssignment/requirements.txt
django>=3.2.14
djangorestframework>=3.13.1
openmeteo-requests>=1.0.0
requests-cache>=0.7.1
retry-requests>=2.0.0
numpy>=1.21.3
pandas>=1.3.3